refactor: Rename expenses component directory into transactions
