+++ packages/sdk-react/src/components/index.ts
@@ -3,7 +3,7 @@ export * from './approvalPolicies';
 export * from './approvalRequests';
 export * from './payables';
 export * from './receivables';
-export * from './expenses';
+export * from './transactions';
 export * from './onboarding';
 export * from './products';
 export * from './tags';
