+++ packages/sdk-react/src/components/transactions/ReceiptDetails/ReceiptDetails.tsx
