+++ packages/sdk-react/src/components/transactions/ReceiptDetails/index.ts
