+++ packages/sdk-react/src/components/transactions/ReceiptPreview/ReceiptPreview.tsx
