+++ packages/sdk-react/src/components/transactions/ReceiptPreview/index.ts
