+++ packages/sdk-react/src/components/transactions/ReceiptsInbox/ReceiptCard.tsx
