+++ packages/sdk-react/src/components/transactions/ReceiptsInbox/ReceiptsInbox.tsx
