+++ packages/sdk-react/src/components/transactions/ReceiptsInbox/index.ts
