+++ packages/sdk-react/src/components/transactions/TransactionDetails/TransactionDetails.tsx
