+++ packages/sdk-react/src/components/transactions/TransactionDetails/index.ts
